Text Cleaner for Vector Corpus
This is a text cleaner for an input in the form of a vector corpus. Notice that you must convert to a plain text document on more than one occasion with the new tm_map update. Built in is also a profanity filter. See the attached text file called bad-words.txt to read in that word list.